feat: add backend tests and configure pytest (#10)
### TL;DR

Added unit tests for backend routes and configured pytest for asyncio. Part of MEE-44.

### What changed?

- Added new test files: `audio_test.py`, `bots_test.py`, and `meetings_test.py`.
- Updated `requirements.txt` with new dependencies for testing.
- Created `conftest.py` to configure pytest to use asyncio as the backend.

### How to test?

In the next PR in the stack I am adding instructions for how to run the tests to the README.

### Why make this change?

This change introduces unit tests for the backend routes, improving code reliability and making it easier to catch potential issues early in the development process.

This is very important as it is a requirement for our capstone, and will additionally increase confidence in the reliability of our service for potential users.

### Links

- FastAPI Async Tests Docs: https://fastapi.tiangolo.com/advanced/async-tests/#example